Save Images into zip file
